utils_projects.py split in two to avoid webserver dependencies
