+++ packages/pytest-simcore/src/pytest_simcore/helpers/utils_projects.py
@@ -6,8 +6,8 @@
 import json
 import re
 import uuid as uuidlib
-from copy import deepcopy
-from typing import Any, Dict, Tuple
+from datetime import datetime
+from typing import Dict
 
 from aiohttp import web
 from simcore_service_webserver.projects.projects_db import (
@@ -15,7 +15,14 @@ from simcore_service_webserver.projects.projects_db import (
     DB_EXCLUSIVE_COLUMNS,
 )
 from simcore_service_webserver.resources import resources
-from simcore_service_webserver.utils import now_str
+
+# FIXME: remove dependencies with simcore_service_webserver
+# TODO:  and merge with utils_projects_extra.py
+
+
+def now_str() -> str:
+    return "{}Z".format(datetime.utcnow().isoformat(timespec="milliseconds"))
+
 
 fake_template_resources = [
     "data/" + name
@@ -91,41 +98,6 @@ async def delete_all_projects(app: web.Application):
     async with db.engine.acquire() as conn:
         query = projects.delete()
         await conn.execute(query)
-
-
-def clone_project_data(project: Dict[str, Any]) -> Tuple[Dict, Dict]:
-    project_copy = deepcopy(project)
-
-    # Update project id
-    # NOTE: this can be re-assigned by dbapi if not unique
-    project_copy_uuid = uuidlib.uuid1()  # random project id
-    project_copy["uuid"] = str(project_copy_uuid)
-
-    # Workbench nodes shall be unique within the project context
-    def _create_new_node_uuid(old_uuid):
-        return str(uuidlib.uuid5(project_copy_uuid, str(old_uuid)))
-
-    nodes_map = {}
-    for node_uuid in project.get("workbench", {}).keys():
-        nodes_map[node_uuid] = _create_new_node_uuid(node_uuid)
-
-    def _replace_uuids(node):
-        if isinstance(node, str):
-            node = nodes_map.get(node, node)
-        elif isinstance(node, list):
-            node = [_replace_uuids(item) for item in node]
-        elif isinstance(node, dict):
-            _frozen_items = tuple(node.items())
-            for key, value in _frozen_items:
-                if key in nodes_map:
-                    new_key = nodes_map[key]
-                    node[new_key] = node.pop(key)
-                    key = new_key
-                node[key] = _replace_uuids(value)
-        return node
-
-    project_copy["workbench"] = _replace_uuids(project_copy.get("workbench", {}))
-    return project_copy, nodes_map
 
 
 class NewProject:
